refactor: task 서비스 계층 추출 — router 1,449줄 → 245줄, SSE 분리
- task/service.py 신설: 상태 조회/모니터링/삭제/취소/로그/DAG 구조 등
  전 엔드포인트 로직 추출 (응답 형태·상태코드·에러 메시지 불변)
- task/sse.py 신설: SSE 스트리밍 루프 분리 (1시간 타임아웃·종료조건·
  finally 정리 동작 동일), router는 EventSourceResponse 래핑만
- Celery 디스패치 kwargs를 TaskSubmission pydantic 스키마로 타입화 —
  .dict() 결과가 기존 kwargs와 동일함을 왕복 테스트로 검증 (와이어 포맷
  불변, 구워커/신웹 혼재 배포 호환)
- tests/unit/test_task_service.py 신규 28건
- 기존 테스트 patch 네임스페이스 갱신 (app.task.sse/app.task.service) —
  SSE 테스트의 stale patch가 실제 폴링 루프를 돌려 테스트 파일 실행이
  62분까지 늘어나던 문제 해결 (2분 46초로 복구, 기대값 불변)

과도기: service의 HTTPException 및 app.main 지연 import(get_celery_app,
테스트 patch 호환)는 PR-8에서 정리
